Added NewSchoolSystem
I have decided to start from scratch on the new school system. I will be copy and pasting certain elements, but I just feel that it would be easier to start completely again, rather than try to reconfigure the entire structure of the previous. I will eventually be removing and deleting the older project, but it will stay during development.